Avoid repeated rank calculations
Modification of the 'elim_and_recalc' and the main 'STRIKE_GOLDD' scripts, in order to avoid unnecessary calculations of the observability rank. In the previous version this could happen when analysing a model with unmeasured inputs ('w'); when adding new Lie derivatives and recalculating the rank for each variable (parameter/state/input), the knowledge about which variables had already been classified as observable was not exploited.